+++ lib/smartest/init_generator.rb
@@ -13,6 +13,66 @@ module Smartest
         Dir[File.join(__dir__, "fixtures", "**", "*.rb")].sort.each do |fixture_file|
           require fixture_file
         end
+
+        Dir[File.join(__dir__, "matchers", "**", "*.rb")].sort.each do |matcher_file|
+          require matcher_file
+        end
+
+        use_matcher PredicateMatcher
+      RUBY
+      "smartest/matchers/predicate_matcher.rb" => <<~RUBY,
+        # frozen_string_literal: true
+
+        module PredicateMatcher
+          def method_missing(name, *arguments, &block)
+            matcher_name = name.to_s
+            return super unless matcher_name.match?(/\\Abe_.+\\z/)
+
+            Matcher.new(matcher_name.delete_prefix("be_"), arguments, block)
+          end
+
+          def respond_to_missing?(name, include_private = false)
+            name.to_s.match?(/\\Abe_.+\\z/) || super
+          end
+
+          class Matcher
+            def initialize(predicate_name, arguments, block)
+              @predicate_name = predicate_name
+              @predicate = "\#{predicate_name}?"
+              @arguments = arguments
+              @block = block
+            end
+
+            def matches?(actual)
+              @actual = actual
+              return false unless actual.respond_to?(@predicate)
+
+              !!actual.public_send(@predicate, *@arguments, &@block)
+            end
+
+            def failure_message
+              return "expected \#{@actual.inspect} to respond to \#{@predicate}" unless @actual.respond_to?(@predicate)
+
+              "expected \#{@actual.inspect} to be \#{description}"
+            end
+
+            def negated_failure_message
+              "expected \#{@actual.inspect} not to be \#{description}"
+            end
+
+            private
+
+            def description
+              return @predicate_name if @arguments.empty?
+
+              "\#{@predicate_name} \#{argument_description}"
+            end
+
+            def argument_description
+              @arguments.map(&:inspect).join(", ")
+            end
+          end
+        end
       RUBY
       "smartest/example_test.rb" => <<~RUBY
         # frozen_string_literal: true
@@ -33,6 +93,7 @@ module Smartest
     def run
       create_directory("smartest")
       create_directory("smartest/fixtures")
+      create_directory("smartest/matchers")
       FILES.each { |path, contents| create_file(path, contents) }
 
       @output.puts